requirements.txt
flask
mysql-connector-python
pandas
yfinance
python-dotenv